refactoring lazy-components
